requirements.txt
base58==1.0.3
Click==7.0
PyYAML==5.1
validate-email==1.3
